Move the AWSProvider object into a Domains module
The object is specifically for finding manipulating AWS `Domain` models
and thus is better suited to be in a Domains module.

This will allow for a `Machines` module to be added at a future date.

Also the class can be called `AWS` as aws-sdk is called `Aws`.
